Updated home screen, adjusting time according to timezone

+++ app/src/main/java/com/jestrabikr/weather/ui/home/HomeUiState.kt
@@ -4,7 +4,7 @@ data class HomeUiState(
     val city: String = "",
     val isLoading: Boolean = false,
     val error: String? = null,
-    val temperature: Double? = null,
+    val temperature: Int? = null,
     val description: String? = null,
     val weatherIcon: String? = null,
     val name: String = ""
